requirements.txt
flask
requests
mysql-connector-python
flask-cors